refactor(filter): extract filter import logic to importers
BREAKING CHANGE: Core filter/feature action import code has been moved
out of the UI controller and into importers. Any controllers extending
a filter import UI must verify they implement getImporter appropriately
and potentially move extended methods to an importer class.
