+++ src/os/filter/im/osfilterimport.js
@@ -1,10 +1,8 @@
 goog.provide('os.filter.im.OSFilterImportCtrl');
 goog.provide('os.filter.im.osFilterImportDirective');
 
-goog.require('os.color');
-goog.require('os.data.BaseDescriptor');
+goog.require('os.filter.im.OSFilterImporter');
 goog.require('os.implements');
-goog.require('os.layer.ILayer');
 goog.require('os.ui.Module');
 goog.require('os.ui.filter.im.FilterImportCtrl');
 
@@ -45,6 +43,15 @@ goog.inherits(os.filter.im.OSFilterImportCtrl, os.ui.filter.im.FilterImportCtrl)
 
 
 /**
+ * @inheritDoc
+ */
+os.filter.im.OSFilterImportCtrl.prototype.getImporter = function() {
+  var layerId = /** @type {string|undefined} */ (this.scope['layerId']);
+  return new os.filter.im.OSFilterImporter(this.getParser(), layerId);
+};
+
+
+/**
  * Get the base filterable descriptors and add in the filterable layers.
  * @inheritDoc
  */
@@ -65,35 +72,4 @@ os.filter.im.OSFilterImportCtrl.prototype.getFilterables = function() {
   }
 
   return filterables;
-};
-
-
-/**
- * @inheritDoc
- */
-os.filter.im.OSFilterImportCtrl.prototype.getProviderFromFilterable = function(filterable) {
-  var provider = os.filter.im.OSFilterImportCtrl.base(this, 'getProviderFromFilterable', filterable);
-
-  // if the filterable implements a provider name interface function, add that to the title
-  if (!provider && os.implements(filterable, os.layer.ILayer.ID)) {
-    provider = /** @type {!os.layer.ILayer} */ (filterable).getProvider();
-  }
-
-  return provider;
-};
-
-
-/**
- * @inheritDoc
- */
-os.filter.im.OSFilterImportCtrl.prototype.getIconsFromFilterable = function(filterable) {
-  if (os.implements(filterable, os.layer.ILayer.ID)) {
-    var options = /** @type {!os.layer.ILayer} */ (filterable).getLayerOptions();
-    var color = /** @type {string|undefined} */ (options['color']);
-    if (color) {
-      return '<i class="fa fa-bars" style="color:' + os.color.toHexString(color) + '"></i>';
-    }
-  }
-
-  return os.filter.im.OSFilterImportCtrl.base(this, 'getIconsFromFilterable', filterable);
 };
